[DEV] SEOS-3020: Change the way qemu_app_wrapper is initialized
Because some platforms require different parameters (e.g. zynqmp) that are not present during test runs on other targets it is better to avoid initializing all QEMU machine configurations on every test run. QemuMachineCfg is a wrapper class that contains a QEMU machine constructor function and a list of parameters and a function to only initialize a single QEMU configuration needed for the current test run